+++ Js2IL.Tests/Import/JavaScript/Import_DynamicImport_Cjs_Namespace_Lib.cjs
@@ -1,0 +1,12 @@
+"use strict";
+
+let value = 1;
+
+module.exports = {
+    inc() {
+        value = value + 1;
+    },
+    get value() {
+        return value;
+    }
+};